# docs/api-spec.yml
# Especificación de API — Sistema de Consulta de Títulos del MEP
# Versión: 0.2.0 (jerarquía-geográfica)
# Formato: OpenAPI 3.0

openapi: "3.0.0"
info:
  title: API de Consulta de Títulos del MEP
  description: API del Sistema de Consulta de Títulos del Ministerio de Educación Pública de Costa Rica
  version: "0.2.0"
servers:
  - url: /
    description: Mismo dominio que la aplicación (Next.js)
paths:
  /api/regiones:
    get:
      summary: Listar todas las regiones activas
      tags: [Regiones]
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Lista de regiones
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Region"
        "401":
          $ref: "#/components/responses/NoAutorizado"
    post:
      summary: Crear una nueva región
      tags: [Regiones]
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [nombre]
              properties:
                nombre:
                  type: string
      responses:
        "201":
          description: Región creada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Region"
        "400":
          $ref: "#/components/responses/ErrorValidacion"
        "401":
          $ref: "#/components/responses/NoAutorizado"
        "403":
          $ref: "#/components/responses/SinPermisos"
  /api/regiones/{id}:
    patch:
      summary: Actualizar una región
      tags: [Regiones]
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [nombre]
              properties:
                nombre:
                  type: string
      responses:
        "200":
          description: Región actualizada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Region"
        "400":
          $ref: "#/components/responses/ErrorValidacion"
        "401":
          $ref: "#/components/responses/NoAutorizado"
        "403":
          $ref: "#/components/responses/SinPermisos"
        "404":
          $ref: "#/components/responses/NoEncontrado"
    delete:
      summary: Desactivar una región (borrado lógico)
      tags: [Regiones]
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Región desactivada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Region"
        "401":
          $ref: "#/components/responses/NoAutorizado"
        "403":
          $ref: "#/components/responses/SinPermisos"
        "404":
          $ref: "#/components/responses/NoEncontrado"
        "409":
          $ref: "#/components/responses/Conflicto"
  /api/escuelas:
    get:
      summary: Listar escuelas activas
      tags: [Escuelas]
      security:
        - cookieAuth: []
      parameters:
        - name: region_id
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Lista de escuelas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Escuela"
        "401":
          $ref: "#/components/responses/NoAutorizado"
    post:
      summary: Crear una nueva escuela
      tags: [Escuelas]
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [regionId, codigoMep, nombre]
              properties:
                regionId:
                  type: integer
                codigoMep:
                  type: string
                nombre:
                  type: string
      responses:
        "201":
          description: Escuela creada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Escuela"
        "400":
          $ref: "#/components/responses/ErrorValidacion"
        "401":
          $ref: "#/components/responses/NoAutorizado"
        "403":
          $ref: "#/components/responses/SinPermisos"
  /api/escuelas/{id}:
    patch:
      summary: Actualizar una escuela
      tags: [Escuelas]
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nombre:
                  type: string
                codigoMep:
                  type: string
      responses:
        "200":
          description: Escuela actualizada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Escuela"
        "400":
          $ref: "#/components/responses/ErrorValidacion"
        "401":
          $ref: "#/components/responses/NoAutorizado"
        "403":
          $ref: "#/components/responses/SinPermisos"
        "404":
          $ref: "#/components/responses/NoEncontrado"
    delete:
      summary: Desactivar una escuela (borrado lógico)
      tags: [Escuelas]
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Escuela desactivada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Escuela"
        "401":
          $ref: "#/components/responses/NoAutorizado"
        "403":
          $ref: "#/components/responses/SinPermisos"
        "404":
          $ref: "#/components/responses/NoEncontrado"
        "409":
          $ref: "#/components/responses/Conflicto"

components:
  schemas:
    Region:
      type: object
      properties:
        id:
          type: integer
        nombre:
          type: string
        activo:
          type: boolean
    Escuela:
      type: object
      properties:
        id:
          type: integer
        regionId:
          type: integer
        codigoMep:
          type: string
        nombre:
          type: string
        activo:
          type: boolean
    Error:
      type: object
      properties:
        error:
          type: string
  responses:
    NoAutorizado:
      description: No autenticado
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    SinPermisos:
      description: No autorizado (rol insuficiente)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    ErrorValidacion:
      description: Error de validación
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NoEncontrado:
      description: Recurso no encontrado
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Conflicto:
      description: Conflicto (regla de negocio)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: next-auth.session-token